requirements.txt
customtkinter
Pillow
opencv-python
pywin32; platform_system == "Windows"
